Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
requests==2.31.0
PyMuPDF==1.23.8
google-generativeai==0.3.2